refactor(web): ModelConfig Code Splitting
